Add support for actors as apps and running unit tests
Apps can now be created as actors. This means that the full force of
the actor system can now be brought to bear on problems. This will
allow things like breaking up work via behaviors, which in turn will
allow GC to run.

Also, you can now supply a TestList object to the application runner
and it will run the unit tests in the specified list if no other
arguments are passed on the command line. This will make it easier to
test parts of the solution.